==> 03_Python Pandas - Tratando e analisando dados/06_criando_agrupamentos.py <==
''' Relatórios de Análise VII
Criando Agrupamentos '''

import pandas as pd 

dados = pd.read_csv('dados\\aluguel_residencial_3.csv', sep=';')
print(dados.head(10))

# Filtrando todos os bairros da nossa base
# dados = dados['Bairro'].unique()

# Criando uma lista com todos os bairros disponíveis
bairros = ['Copacabana', 'Jardim Botânico', 'Centro', 'Higienópolis', 'Cachambi'
 'Barra da Tijuca', 'Ramos', 'Grajaú', 'Lins de Vasconcelos', 'Taquara'
 'Freguesia (Jacarepaguá)', 'Tijuca', 'Olaria', 'Ipanema', 'Campo Grande'
 'Botafogo', 'Recreio dos Bandeirantes', 'Leblon', 'Jardim Oceânico'
 'Humaitá', 'Península', 'Méier', 'Vargem Pequena', 'Maracanã', 'Jacarepaguá'
 'São Conrado', 'Vila Valqueire', 'Gávea', 'Cosme Velho', 'Bonsucesso'
 'Todos os Santos', 'Laranjeiras', 'Itanhangá', 'Flamengo', 'Piedade', 'Lagoa'
 'Catete', 'Jardim Carioca', 'Benfica', 'Glória', 'Praça Seca', 'Vila Isabel'
 'Engenho Novo', 'Engenho de Dentro', 'Pilares', 'Água Santa', 'São Cristóvão'
 'Ilha do Governador', 'Jardim Sulacap', 'Oswaldo Cruz', 'Vila da Penha'
 'Anil', 'Vargem Grande', 'Tanque', 'Vaz Lobo', 'Madureira'
 'São Francisco Xavier', 'Pechincha', 'Leme', 'Irajá', 'Quintino Bocaiúva'
 'Urca', 'Penha', 'Gardênia Azul', 'Rio Comprido', 'Andaraí', 'Santa Teresa'
 'Inhaúma', 'Marechal Hermes', 'Curicica', 'Santíssimo', 'Moneró', 'Camorim'
 'Cascadura', 'Praia da Bandeira', 'Saúde', 'Joá', 'Realengo', 'Fátima'
 'Inhoaíba', 'Rocha', 'Jardim Guanabara', 'Jabour', 'Braz de Pina'
 'Praça da Bandeira', 'Vila Kosmos', 'Vista Alegre', 'Encantado', 'Campinho'
 'Guaratiba', 'Riachuelo', 'Bangu', 'Lapa', 'Catumbi', 'Penha Circular'
 'Abolição', 'Tomás Coelho', 'Colégio', 'Pavuna', 'Santa Cruz'
 'Alto da Boa Vista', 'Cidade Nova', 'Bento Ribeiro', 'Estácio'
 'Jardim América', 'Cordovil', 'Caju', 'Pedra de Guaratiba', 'Padre Miguel'
 'Paciência', 'Del Castilho', 'Arpoador', 'Sampaio', 'Anchieta', 'Icaraí'
 'Senador Vasconcelos', 'Rocha Miranda', 'Gamboa', 'Maria da Graça'
 'Barra de Guaratiba', 'Vicente de Carvalho', 'Paquetá', 'Largo do Machado'
 'Parada de Lucas', 'Freguesia (Ilha do Governador)', 'Portuguesa'
 'Guadalupe', 'Parque Anchieta', 'Turiaçu', 'Pitangueiras', 'Vila Militar'
 'Vidigal', 'Senador Camará', 'Usina', 'Vigário Geral', 'Cosmos', 'Jacaré'
 'Cocotá', 'Honório Gurgel', 'Engenho da Rainha', 'Cachamorra', 'Zumbi', 'Tauá'
 'Santo Cristo', 'Ribeira', 'Magalhães Bastos', 'Cacuia', 'Bancários'
 'Cavalcanti', 'Rio da Prata', 'Cidade Jardim', 'Coelho Neto']


# Realizando uma seleção e criando um novo DataFrame
selecao = dados['Bairro'].isin(bairros)                 # Criamos uma seleção onde a minha lista tem que ser igual a minha coluna Bairro
dados = dados[selecao]                                  # Criamos um novo DataFrame
sem_duplicadas = dados['Bairro'].drop_duplicates()      # Removemos as duplicadas com o drop_duplicates
print(sem_duplicadas)

# Criamos um grupo com os dados 
grupo_bairro = dados.groupby('Bairro')                  # Criamos um grupo com os valores da coluna Bairro
v = (grupo_bairro[['Valor', 'Condominio']].mean()).round(2)             # No meu grupo, eu peço o Valor de cada item e passo o metodo mean para apresentar a media deles
print(v)

''' Estatísticas Descritivas '''
# Utilizando o describe para pegar estatisticas
b = grupo_bairro['Valor'].describe().round(2)
print(b)

# Utilizando o aggregate para selecionar quais estatisticas queremos plotar
b = grupo_bairro['Valor'].aggregate(['min','max','sum']).rename(columns = {'min':'Mínimo', 'max':'Máximo', 'sum':'Soma'})
print(b)



import matplotlib.pyplot as plt
plt.rc('figure', figsize = (7,7))     # Configura o tamanho que a figura irá ser apresentada

# Apresentando os valores de aluguel por bairro com o matplotlib
fig = grupo_bairro['Valor'].mean().plot.bar(color = 'blue')
fig = grupo_bairro['Condominio'].mean().plot.bar(color = 'red')
fig.set_ylabel('Valor do Aluguel x Condominio')
fig.set_xlabel('Bairro')
fig.set_title("Valor Médio do Aluguel x Condominio por Bairro", {'fontsize':20})
plt.show()
